libdrm: Avoid applying twice the same patch
Mask in local.conf the libdrm recipe from BSP layer to
avoid applying the drm-update-arm.patch which is already
applied from meta-freescale layer.
We also add in our layers a libdrm bbappend file to set the
SRCREV of the package as in the BSP layer.

Changelog-entry: libdrm: Avoid applying twice the same patch
Signed-off-by: Sebastian Panceac <sebastian@balena.io>