linux-variscite: Build net/dummy as module
Add dummy net driver support, which is used to
check if a container is privileged during
udev initialization.

Building it as module ensures that dummy
devices are not needlessly created during boot.

This commit is taken from Alex's PR
https://github.com/balena-os/meta-balena/pull/1703/commits
so we can drop these changes here when a meta-balena with
that PR is released and we will use it.

Changelog-entry: Build net/dummy as module
Signed-off-by: Florin Sarbu <florin@balena.io>