recipes-bsp/u-boot: Fix 6UL not booting, disable bootswitch check
We increase the load address for the kernel to be above the dtb
address to fix booting, and also disable the bootswitch check
to allow for autokit provisioning. We now do the check by
iterating through mmc0 (sd) and mmc1 (eMMC) and set the
dtb based on that. Previously, the boot switch dictated
which dtb is loaded - either sd-card without wifi or eMMC with wifi
and no sd-card support.

Signed-off-by: Alexandru Costache <alexandru@balena.io>