imx6ul-var-dart.yml: Remove automated testing
With the Scarthgap update the SPL won't run (we have a VSM-6UL-739A SoM) and because of this meta-balena PRs are blocked.

We have opened a ticket with Variscite to investigate (ticket id 26934 in https://variscite.com/support/)
but until we have a fix we won't do automatic releases for this board.

Signed-off-by: Florin Sarbu <florin@balena.io>